move sharding logic from GroupedParameterBuilder to ShardingSQLRewriteEngine
